automatically backed up by learn

+++ index.js
@@ -30,5 +30,5 @@ function dec(b) {
 }
 
 function makeInt(a) {
-  return parseInt('a', 10)
+  return parseFloat('a', 10)
 }
